Phase B-1: kx-model-validator lib.rs split (Rule 3, behavior-preserving move-only)
First Phase B refactor PR following the Engineering Discipline / Rule 3
recipe (no giant lib.rs). Pure move; zero behavior change.

Before: src/lib.rs 1187 lines (capability primitives + RequiredCapabilities +
ProvidedCapabilities + outcome vocabulary + check fn + ModelRegistry +
Recommender + 18 inline unit tests, all in one file).

After: thin src/lib.rs (95 lines: crate attrs + crate doc + mod decls +
pub use re-exports + #[cfg(test)] mod tests;) + 8 sibling modules:
- capabilities.rs (Modality, Quantization, License, LicenseConstraint +
  private commercial_use_permitted / redistribution_permitted helpers)
- requirements.rs (RequiredCapabilities + permissive() constructor)
- provided.rs (ProvidedCapabilities + Soundness + builder impls)
- outcome.rs (DegradationReason + MissingCapability + ValidatorOutcome)
- check.rs (the pure/total/deterministic check fn)
- registry.rs (ModelRegistry trait + InMemoryModelRegistry)
- recommender.rs (RankingPolicy + Candidate + Recommender +
  private outcome_rank helper)
- tests.rs (the 18 inline unit tests; #[cfg(test)] from lib.rs)

Re-exports preserve every previously-public path — `kx_model_validator::Modality`,
`::check`, `::Recommender`, etc. all unchanged. No downstream crate's `use`
statements needed editing (verified by `cargo build --workspace` clean).

Local gates green on Apple Silicon (--all-features):
- cargo build -p kx-model-validator: clean
- cargo test --workspace --all-features: 645 passed + 8 ignored (IDENTICAL
  to pre-split baseline of 645/8)
- cargo test -p kx-model-validator: 18 + 11 + 5 = 34 (identical)
- cargo clippy -p kx-model-validator --all-targets -- -D warnings: clean
- cargo fmt --all -- --check: clean
- RUSTDOCFLAGS=-D warnings cargo doc -p kx-model-validator --no-deps: clean

Next Phase B slice: kx-tool-registry split (1504 → thin lib.rs + siblings).
